operationize find and aggregate
